added scripts directory and template file
