+++ scripts/read_nft_id.cdc
